Refactor query tag tests
