+++ tests/integration/query_tag_tests/models/incremental_model_query_tag.sql
@@ -1,0 +1,7 @@
+{{ config(
+    materialized = 'incremental',
+    unique_key = 'id',
+    query_tag = var("query_tag") + '_incremental'
+) }}
+
+select 1 as id
