+++ tests/integration/query_tag_tests/models/table_model_query_tag.sql
@@ -1,0 +1,6 @@
+{{ config(
+    materialized = 'table',
+    query_tag = var("query_tag") + '_table'
+) }}
+
+select 1 as id
