+++ tests/integration/query_tag_tests/models/view_model_query_tag.sql
@@ -1,0 +1,6 @@
+{{ config(
+    materialized = 'view',
+    query_tag = var("query_tag") + '_view'
+) }}
+
+select 1 as id
