+++ tests/integration/query_tag_tests/seeds/seed_query_tag.csv
@@ -1,0 +1,2 @@
+id
+1
